Use fixed-width integers in Tiberian Dawn
Migrate the remaining `long` and `unsigned long` in the Tiberian Dawn
sources:

- Radio messages carry their parameter as int32_t through the whole
  Receive_Message chain and LParam, as in Red Alert.
- Game and name CRCs are uint32_t. They were 64-bit on LP64, but only
  the low 32 bits ever mattered and EventLength is taken from sizeof,
  so the frame-info event is 4 bytes again, as in the original protocol.
- Prerequisite masks are uint64_t like Red Alert's; the house's 32-bit
  truncation of the mask is unchanged.
- Spare ticks, credits and timer arithmetic keep 64 bits; trigger house
  events take their data as int64_t so credit checks do not narrow.
- Obfuscate returns the uint32_t code its keyphrases are compared with.
- Key frame offsets stored in pointers are read with std::bit_cast at
  pointer width, and the pathfinding cross product computes in int64_t.

No serialized member changes width, so the save version is unchanged.

+++ src/td/aircraft.cc
@@ -2508,7 +2508,7 @@ COORDINATE AircraftClass::Target_Coord() const {
  *=============================================================================================*/
 RadioMessageType AircraftClass::Receive_Message(RadioClass* from,
                                                 RadioMessageType message,
-                                                long& param) {
+                                                int32_t& param) {
   Validate();
   switch (message) {
     case RADIO_PREPARED:
@@ -2610,7 +2610,7 @@ RadioMessageType AircraftClass::Receive_Message(RadioClass* from,
             if (cell == 0) {
               Transmit_Message(RADIO_OVER_OUT, from);
             } else {
-              param = static_cast<long>(::As_Target(cell));
+              param = static_cast<int32_t>(::As_Target(cell));
 
               /*
               **	Tell the potential passenger where it should go. If the
@@ -2619,7 +2619,7 @@ RadioMessageType AircraftClass::Receive_Message(RadioClass* from,
               */
               if (Transmit_Message(RADIO_MOVE_HERE, param, from) ==
                   RADIO_YEA_NOW_WHAT) {
-                param = static_cast<long>(As_Target());
+                param = static_cast<int32_t>(As_Target());
                 Transmit_Message(RADIO_TETHER);
                 if (Transmit_Message(RADIO_MOVE_HERE, param, from) !=
                     RADIO_ROGER) {
