+++ src/td/map.cc
@@ -885,8 +885,8 @@ void MapClass::Overlap_Up(CELL cell, ObjectClass* object) {
  *tiberium worth.                                            * 02/15/1995 JLB :
  *Optimal scan.                                                            *
  *=============================================================================================*/
-long MapClass::Overpass() {
-  long value = 0;
+int32_t MapClass::Overpass() {
+  int32_t value = 0;
 
   /*
   **	Smooth out Tiberium. Cells that are not surrounded by other tiberium
@@ -917,7 +917,7 @@ long MapClass::Overpass() {
 #ifdef DEMO
 bool MapClass::Read_Binary(const char* root, unsigned long*)
 #else
-bool MapClass::Read_Binary(const char* root, unsigned long* crc)
+bool MapClass::Read_Binary(const char* root, uint32_t* crc)
 #endif
 {
   CCFileClass file;
